Disable VectorTest on Z
Z12 and lower architectures don't have vector support. This causes
illegal instruction failures on Z. Since we cannot currently
determine at runtime which Z architecture the test is running on, the
tril test for VectorTest is disabled on Z altogether. We can re-enable
once we are able to check at runtime.

Closes: #1941

Signed-off-by: Dhruv Chopra <Dhruv.C.Chopra@ibm.com>