Correct SPDX license identifier
The original SPDX license identifier was missing the refined details
about the GPL-2.0 special clauses. The new identifier includes these
details explicitly and more precisely depicts the Eclipse OMR license.

For every file in the repository, this commit changes this line in
the file header :
SPDX-License-Identifier: EPL-2.0 OR Apache-2.0

to :
SPDX-License-Identifier: EPL-2.0 OR Apache-2.0 OR GPL-2.0 WITH Classpath-exception-2.0 OR LicenseRef-GPL-2.0 WITH Assembly-exception

Fixes: #2172

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>